refactor: change architecture clean and update naming
