+++ src/commands/inline.ts
@@ -1,26 +1,25 @@
 import { Composer, Context } from "grammy";
-import { Pacman } from "../types/Pacman";
-import { Tealdeer } from "../types/Tealdeer";
+import { PacmanSearch, TealdeerSearch } from "../services";
 
 const composer = new Composer();
 
 composer.inlineQuery(/(.*)/ig, async (ctx: Context) => {
   let search: string | undefined;
-  let instance: Pacman | Tealdeer = new Pacman();
+  let instance: PacmanSearch | TealdeerSearch = new PacmanSearch();
 
   if (!ctx.inlineQuery?.query) {
-    return await ctx.answerInlineQuery(await instance.noQuery());
+    return await ctx.answerInlineQuery(await instance.getEmptyQuery());
   }
 
   const split = ctx.inlineQuery?.query.split(" ");
 
   switch (split![0]) {
     case "tldr":
-      instance = new Tealdeer();
+      instance = new TealdeerSearch();
       search = split?.slice(1).join(" ");
       break;
     default:
-      instance = new Pacman();
+      instance = new PacmanSearch();
       search = split?.join(" ");
       break;
   }
@@ -28,11 +27,11 @@ composer.inlineQuery(/(.*)/ig, async (ctx: Context) => {
   await instance.search(search);
 
   if (instance.getLength() === 0) {
-    return await ctx.answerInlineQuery(await instance.notFound(search));
+    return await ctx.answerInlineQuery(await instance.getNotFound(search));
   }
 
   return await ctx.answerInlineQuery(
-    await instance.inline(),
+    await instance.getResults(),
   );
 });
