+++ src/commands/text.ts
