+++ src/config/index.ts
@@ -1,0 +1,3 @@
+export { default as Config, BotConfig } from "./config";
+export { default as parseArgs } from "./cli";
+
