+++ src/services/index.ts
@@ -1,0 +1,3 @@
+export { PacmanSearch, TealdeerSearch, SearchService } from "./package-search";
+export type { Package } from "./package-search";
+
